issue with parsing an int into the price and weight needs fixing

+++ app/src/main/java/com/example/fruitapp/FruitInfo.kt
@@ -1,3 +1,3 @@
 package com.example.fruitapp
 
-data class FruitInfo(val type: String, val price: Float, val weight: Float)
+data class FruitInfo(val type: String, val price: Int, val weight: Int)
